LFS-1022: Remove Progeny/Ancestors
