+++ modules/subjects/src/main/java/ca/sickkids/ccm/lfs/subjects/internal/SubjectParentEditor.java
@@ -19,6 +19,7 @@ package ca.sickkids.ccm.lfs.subjects.internal;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.jcr.Node;
 import javax.jcr.RepositoryException;
 import javax.jcr.Session;
 
@@ -32,16 +33,16 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * An {@link Editor} that sets the {@code progeny} property for every {@code lfs:Subject}
- * or {@code lfs:SubjectType} with the children of that node.
+ * An {@link Editor} that sets the {@code parent} property for every {@code lfs:Subject}
+ * or {@code lfs:SubjectType} with the direct parent of that node.
  *
  * @version $Id$
  */
-public class SubjectProgenyEditor extends DefaultEditor
+public class SubjectParentEditor extends DefaultEditor
 {
-    private static final Logger LOGGER = LoggerFactory.getLogger(SubjectProgenyEditor.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(SubjectParentEditor.class);
 
-    private static final String PROP_PROGENY = "progeny";
+    private static final String PROP_PARENTS = "parents";
 
     private final Session session;
 
@@ -55,7 +56,7 @@ public class SubjectProgenyEditor extends DefaultEditor
      * @param nodeBuilder the current node
      * @param session the session used to retrieve subjects by UUID
      */
-    public SubjectProgenyEditor(final NodeBuilder nodeBuilder, final Session session)
+    public SubjectParentEditor(final NodeBuilder nodeBuilder, final Session session)
     {
         this.currentNodeBuilder = nodeBuilder;
         this.session = session;
@@ -69,13 +70,13 @@ public class SubjectProgenyEditor extends DefaultEditor
     public Editor childNodeAdded(final String name, final NodeState after)
         throws CommitFailedException
     {
-        return new SubjectProgenyEditor(this.currentNodeBuilder.getChildNode(name), this.session);
+        return new SubjectParentEditor(this.currentNodeBuilder.getChildNode(name), this.session);
     }
 
     @Override
     public Editor childNodeChanged(String name, NodeState before, NodeState after) throws CommitFailedException
     {
-        return new SubjectProgenyEditor(this.currentNodeBuilder.getChildNode(name), this.session);
+        return new SubjectParentEditor(this.currentNodeBuilder.getChildNode(name), this.session);
     }
 
     @Override
@@ -83,51 +84,34 @@ public class SubjectProgenyEditor extends DefaultEditor
     {
         if (isSubject(this.currentNodeBuilder)) {
             try {
-                computeProgeny();
+                computeParent();
             } catch (RepositoryException e) {
-                // This is not a fatal error, the subject progeny is not required for a functional application
-                LOGGER.warn("Unexpected exception while computing the progeny of subject {}",
+                // This is not a fatal error, the subject parentage is not required for a functional application
+                LOGGER.warn("Unexpected exception while computing the parentage of subject {}",
                     this.currentNodeBuilder.getString("jcr:uuid"));
             }
         }
     }
 
     /**
-     * Gather all children of a given node and store them in the {@code progeny} property.
+     * Gather the parent of a given node and store it in the {@code parents} property, removing it if there is none.
      *
      * @throws RepositoryException if accessing the repository fails
      */
-    private void computeProgeny() throws RepositoryException
+    private void computeParent() throws RepositoryException
     {
-        final List<String> identifiers = getProgeny(this.currentNodeBuilder);
+        final List<String> identifiers = new ArrayList<>();
+        Node parentNode = this.session.getNodeByIdentifier(this.currentNodeBuilder.getString("jcr:uuid")).getParent();
 
-        // Get&write progeny to the JCR repo
-        // NB: Empty properties cause errors in JCR-SQL2 queries, so we remove the property altogether if it is empty
-        if (identifiers.size() > 0) {
-            this.currentNodeBuilder.setProperty(PROP_PROGENY, identifiers, Type.WEAKREFERENCES);
-        } else if (this.currentNodeBuilder.hasProperty(PROP_PROGENY)) {
-            this.currentNodeBuilder.removeProperty(PROP_PROGENY);
+        // Iterate through all parents of this node
+        if (parentNode != null && parentNode.hasProperty("jcr:primaryType") && isSubject(parentNode)) {
+            this.currentNodeBuilder.setProperty(PROP_PARENTS, parentNode.getProperty("jcr:uuid").getString(),
+                Type.WEAKREFERENCE);
+        } else {
+            if (this.currentNodeBuilder.hasProperty(PROP_PARENTS)) {
+                this.currentNodeBuilder.removeProperty(PROP_PARENTS);
+            }
         }
-    }
-
-    /**
-     * Gather all identifiers from all the Subject/SubjectType children of the current node
-     * and store them as a string list.
-     *
-     * @throws RepositoryException if accessing the repository fails
-     */
-    private List<String> getProgeny(NodeBuilder node) throws RepositoryException
-    {
-        List<String> progeny = new ArrayList<>();
-
-        // Iterate through all children of this node
-        for (String childName : node.getChildNodeNames()) {
-            NodeBuilder childNode = node.getChildNode(childName);
-            progeny.add(childNode.getProperty("jcr:uuid").getValue(Type.STRING));
-            progeny.addAll(getProgeny(childNode));
-        }
-
-        return progeny;
     }
 
     /**
@@ -142,6 +126,17 @@ public class SubjectProgenyEditor extends DefaultEditor
     }
 
     /**
+     * Checks if the given node is a Subject or SubjectType node.
+     *
+     * @param node the JCR Node to check
+     * @return {@code true} if the node is a Subject/SubjectType node, {@code false} otherwise
+     */
+    private boolean isSubject(Node node) throws RepositoryException
+    {
+        return "lfs:Subject".equals(getNodeType(node)) || "lfs:SubjectType".equals(getNodeType(node));
+    }
+
+    /**
      * Retrieves the primary node type of a node, as a String.
      *
      * @param node the node whose type to retrieve
@@ -150,5 +145,16 @@ public class SubjectProgenyEditor extends DefaultEditor
     private String getNodeType(NodeBuilder node)
     {
         return node.getProperty("jcr:primaryType").getValue(Type.STRING);
+    }
+
+    /**
+     * Retrieves the primary node type of a node, as a String.
+     *
+     * @param node the node whose type to retrieve
+     * @return a string
+     */
+    private String getNodeType(Node node) throws RepositoryException
+    {
+        return node.getProperty("jcr:primaryType").getString();
     }
 }
