+++ modules/subjects/src/main/java/ca/sickkids/ccm/lfs/subjects/internal/SubjectParentEditorProvider.java
@@ -32,12 +32,12 @@ import org.osgi.service.component.annotations.ReferenceCardinality;
 import org.osgi.service.component.annotations.ReferencePolicyOption;
 
 /**
- * A {@link EditorProvider} returning {@link SubjectAncestryEditor}.
+ * A {@link EditorProvider} returning {@link SubjectParentEditor}.
  *
  * @version $Id$
  */
 @Component(immediate = true)
-public class SubjectAncestryEditorProvider implements EditorProvider
+public class SubjectParentEditorProvider implements EditorProvider
 {
     @Reference(fieldOption = FieldOption.REPLACE, cardinality = ReferenceCardinality.OPTIONAL,
         policyOption = ReferencePolicyOption.GREEDY)
@@ -50,8 +50,8 @@ public class SubjectAncestryEditorProvider implements EditorProvider
     {
         if (this.rrf != null && this.rrf.getThreadResourceResolver() != null) {
             Session session = this.rrf.getThreadResourceResolver().adaptTo(Session.class);
-            // Each SubjectAncestryEditor maintains a state, so a new instance must be returned each time
-            return new SubjectAncestryEditor(builder, session);
+            // Each SubjectParentEditor maintains a state, so a new instance must be returned each time
+            return new SubjectParentEditor(builder, session);
         }
         return null;
     }
